Criaçao de novo métodos

+++ projeto_POO-BD/src/main/java/com/api/projeto_poobd/repositories/RepertorioRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepertorioRepository extends JpaRepository<RepertorioModel, Long> {
+public interface RepertorioRepository extends JpaRepository<RepertorioModel, Integer> {
+
 }
